Undo user-facing changes for review follows.
The last push to the hg repo was for a project working on
how we might visualize stuff for the JGI. This change rolls
back the user-facing changes so that we can have a single
HEAD on the branch while still continuing development. Some
of the plumbing, including an is_review URL parameter, makes
sense to keep. We just don't want to expose it yet.