revert change and fix: dataTableModel rough parity
Revert change: the dataTableModel used in ODK Survey needs to come from the formDef.json because session variables and default values are only defined in the formDef and not in the model returned by Java.
Fix: the dataTableModel returned by Java, while a complete JSON schema, does not recursively enumerate the fields within objects (complex data structures) so that all of an object's nested structure is easily examined by traversing the fields of this modelling object. Modify odkData to perform this recursive unfolding. This ensures that apps can simply traverse the fields within the dataTableModel and see all data fields that are persisted to the database. Fields that are nested within objects are identified by their elementPath field property not matching their elementKey property .