fix: doAction callback processing confusion
(1) Change the doAction callback implementation to use enqueueTriggeringContext(), which queues and dispatches callback and event ctxts in the order in which they were enqueued.
(2) Add registrationGeneration tracking and guards to ensure that only the callbacks for the active registration can remove a queued action. Add logging when we are executing an incorrect generation.